refactor(scripts): Move combineExceptionFiles to utils.
