+++ packages/scripts/src/__tests__/util.ts
@@ -1,4 +1,5 @@
 import {
+  combineExceptionFiles,
   expandGroupIds,
   shouldProcessFile,
   sortSourceAndYmlFiles
@@ -82,6 +83,30 @@ describe("util", () => {
       expect(idsFromGroups.includes("group2.key3")).toBe(true);
       expect(idsFromGroups.includes("group2.key4")).toBe(true);
       expect(idsFromGroups.includes("group2.key5")).toBe(true);
+    });
+  });
+
+  describe("combineExceptionFiles", () => {
+    it("should combine ignored ids", async () => {
+      const exceptionFiles = [
+        `${mocksFolderFromCwd}/i18n1/i18n-exceptions.json`,
+        `${mocksFolderFromCwd}/i18n2/i18n-exceptions.json`
+      ];
+      const { groups, ignoredIds } = await combineExceptionFiles(
+        exceptionFiles
+      );
+
+      const groupKeys = Object.keys(groups);
+      expect(groupKeys.length).toBe(1);
+      expect(groupKeys[0]).toBe("otpUi.OtherComponent.*Message");
+
+      expect(ignoredIds.size).toBe(2);
+      expect(
+        ignoredIds.has("otpUi.TestComponent1.unusedTextThatIsIgnored")
+      ).toBe(true);
+      expect(
+        ignoredIds.has("otpUi.TestComponent2.unusedTextThatIsIgnored")
+      ).toBe(true);
     });
   });
 });
