+++ packages/scripts/src/validate-i18n.ts
@@ -1,48 +1,14 @@
 /* eslint-disable no-console */
 import flatten from "flat";
-import { promises as fs } from "fs";
 import { extract } from "@formatjs/cli";
 
 import {
+  combineExceptionFiles,
   expandGroupIds,
   isNotSpecialId,
   loadYamlFile,
   sortSourceAndYmlFiles
 } from "./util";
-
-interface CheckException {
-  groups: Record<string, string[]>;
-  ignoredIds: Set<string>;
-}
-
-/**
- * Combines exception files into a single exception object.
- */
-export async function combineExceptionFiles(
-  exceptionFiles: string[]
-): Promise<CheckException> {
-  let allIgnoredIds = [];
-  const allGroups = [];
-  await Promise.all(
-    exceptionFiles.map(async file => {
-      const rawJson = (await fs.readFile(file)).toString();
-      const jsonObject = JSON.parse(rawJson);
-      allIgnoredIds = allIgnoredIds.concat(jsonObject.ignoredIds);
-      if (jsonObject.groups) {
-        allGroups.push(jsonObject.groups);
-      }
-    })
-  );
-  const groups = allGroups.reduce(
-    (result, group) => ({ ...result, ...group }),
-    {}
-  );
-  return {
-    groups,
-    // Make sure ignored ids are unique
-    ignoredIds: new Set(allIgnoredIds)
-  };
-}
 
 /**
  * Computes the unused ids from code or YML file for a given locale.
